move configuration for jobConf to HadoopRDD
