+++ core/src/main/scala/org/apache/spark/rdd/HadoopRDD.scala
@@ -39,7 +39,6 @@ import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.deploy.SparkHadoopUtil
 import org.apache.spark.util.NextIterator
 
-
 /**
  * A Spark split class that wraps around a Hadoop InputSplit.
  */
@@ -117,7 +116,7 @@ class HadoopRDD[K, V](
 
   protected val inputFormatCacheKey = "rdd_%d_input_format".format(id)
 
-  // used to build JT ID
+  // used to build JobTracker ID
   private val createTime = new Date()
 
   // Returns a JobConf that will be used on slaves to obtain input splits for Hadoop reads.
@@ -175,28 +174,13 @@ class HadoopRDD[K, V](
   override def compute(theSplit: Partition, context: TaskContext) = {
     val iter = new NextIterator[(K, V)] {
 
-      private def localizeConfiguration(conf: JobConf) {
-        //generate job id
-        val stageId = context.stageId
-        val dummyJobTrackerID = new SimpleDateFormat("yyyyMMddHHmm").format(createTime)
-        val jobId = new JobID(dummyJobTrackerID, stageId)
-        val splitID = theSplit.index
-        val attemptId = (context.attemptId % Int.MaxValue).toInt
-        val taId = new TaskAttemptID(new TaskID(jobId, true, splitID), attemptId)
-
-        conf.set("mapred.tip.id", taId.getTaskID.toString)
-        conf.set("mapred.task.id", taId.toString)
-        conf.setBoolean("mapred.task.is.map", true)
-        conf.setInt("mapred.task.partition", splitID)
-        conf.set("mapred.job.id", jobId.toString)
-      }
-
       val split = theSplit.asInstanceOf[HadoopPartition]
       logInfo("Input split: " + split.inputSplit)
       var reader: RecordReader[K, V] = null
       val jobConf = getJobConf()
       val inputFormat = getInputFormat(jobConf)
-      localizeConfiguration(jobConf)
+      HadoopRDD.addLocalConfiguration(new SimpleDateFormat("yyyyMMddHHmm").format(createTime),
+        context.stageId, theSplit.index, context.attemptId.toInt, jobConf)
       reader = inputFormat.getRecordReader(split.inputSplit.value, jobConf, Reporter.NULL)
 
       // Register an on-task-completion callback to close the input stream.
@@ -248,4 +232,27 @@ private[spark] object HadoopRDD {
 
   def putCachedMetadata(key: String, value: Any) =
     SparkEnv.get.hadoopJobMetadata.put(key, value)
+
+  /**
+   *
+   * @param jtId
+   * @param jobId
+   * @param splitId
+   * @param attemptId
+   * @param conf
+   */
+  def addLocalConfiguration(jtId: String, jobId: Int, splitId: Int, attemptId: Int,
+                            conf: JobConf) {
+    // generate job id
+    //val stageId = context.stageId
+    val jobID = new JobID(jtId, jobId)
+    //val attemptId = (attemptId % Int.MaxValue).toInt
+    val taId = new TaskAttemptID(new TaskID(jobID, true, splitId), attemptId)
+
+    conf.set("mapred.tip.id", taId.getTaskID.toString)
+    conf.set("mapred.task.id", taId.toString)
+    conf.setBoolean("mapred.task.is.map", true)
+    conf.setInt("mapred.task.partition", splitId)
+    conf.set("mapred.job.id", jobID.toString)
+  }
 }
